+++ packages/@office-iss/react-native-win32/src/Libraries/Core/ReactNativeVersionCheck.win32.js
@@ -40,7 +40,17 @@ exports.checkVersions = function checkVersions(): void {
   }
 };
 
-function _formatVersion(version): string {
+function _formatVersion(
+  version:
+    | {major: number, minor: number, patch: number, prerelease: ?number}
+    | $TEMPORARY$object<{
+        major: number,
+        minor: number,
+        patch: number,
+        prerelease: null,
+      }>
+    | {major: number, minor: number, patch: number, prerelease: ?string},
+): string {
   return (
     `${version.major}.${version.minor}.${version.patch}` +
     // eslint-disable-next-line eqeqeq