+++ packages/@react-native-windows/cli/src/e2etest/projectConfig.utils.ts
@@ -32,14 +32,10 @@ export async function ensureWinUI3Project(folder: string) {
       {
         id: 'Microsoft.ReactNative.Cxx',
         version: '1.0.0',
-        hasProps: false,
-        hasTargets: true,
       },
       {
         id: 'Microsoft.UI.Xaml',
         version: '2.3.4.5',
-        hasProps: false,
-        hasTargets: false,
       },
       {
         id: 'Microsoft.WindowsAppSDK',
@@ -63,13 +59,6 @@ export async function ensureWinUI3Project(folder: string) {
   await copyAndReplace(
     path.join(templateRoot, 'cpp-app/proj/MyApp.vcxproj'),
     path.join(projDir, 'WithWinUI3.vcxproj'),
-    replacements,
-    null
-  );
-
-  await copyAndReplace(
-    path.join(templateRoot, 'cpp-app/proj/packages.config'),
-    path.join(projDir, 'packages.config'),
     replacements,
     null
   );