+++ packages/@react-native-windows/cli/src/generator-common/index.ts
@@ -27,7 +27,7 @@ function walk(current: string): string[] {
 
   const files = fs
     .readdirSync(current)
-    .map(child => walk(path.join(current, child)));
+    .map((child) => walk(path.join(current, child)));
   const result: string[] = [];
   return result.concat.apply([current], files);
 }
@@ -62,11 +62,11 @@ export function resolveContents(
 
   if (replacements.useMustache) {
     content = mustache.render(content, replacements);
-    (replacements.regExpPatternsToRemove || []).forEach(regexPattern => {
+    (replacements.regExpPatternsToRemove || []).forEach((regexPattern) => {
       content = content.replace(new RegExp(regexPattern, 'g'), '');
     });
   } else {
-    Object.keys(replacements).forEach(regex => {
+    Object.keys(replacements).forEach((regex) => {
       content = content.replace(new RegExp(regex, 'g'), replacements[regex]);
     });
   }
@@ -131,7 +131,7 @@ export async function copyAndReplace(
       shouldOverwrite = await contentChangedCallback(destPath, contentChanged);
     }
     if (shouldOverwrite === 'overwrite') {
-      copyBinaryFile(srcPath, destPath, err => {
+      copyBinaryFile(srcPath, destPath, (err) => {
         if (err) {
           throw err;
         }
@@ -181,13 +181,13 @@ function copyBinaryFile(
   let cbCalled = false;
   const srcPermissions = fs.statSync(srcPath).mode;
   const readStream = fs.createReadStream(srcPath);
-  readStream.on('error', err => {
+  readStream.on('error', (err) => {
     done(err);
   });
   const writeStream = fs.createWriteStream(destPath, {
     mode: srcPermissions,
   });
-  writeStream.on('error', err => {
+  writeStream.on('error', (err) => {
     done(err);
   });
   writeStream.on('close', () => {