+++ packages/@react-native-windows/codegen/src/Cli.ts
@@ -98,9 +98,11 @@ function checkFilesForChanges(
 
   const allExistingFiles = globby
     .sync(`${outputDir}/**`)
-    .map(_ => path.normalize(_))
+    .map((_) => path.normalize(_))
     .sort();
-  const allGeneratedFiles = [...map.keys()].map(_ => path.normalize(_)).sort();
+  const allGeneratedFiles = [...map.keys()]
+    .map((_) => path.normalize(_))
+    .sort();
 
   if (
     allExistingFiles.length !== allGeneratedFiles.length ||
@@ -130,7 +132,7 @@ function writeMapToFiles(map: Map<string, string>, outputDir: string) {
 
   // This ensures that we delete any generated files from modules that have been deleted
   const allExistingFiles = globby.sync(`${outputDir}/**`);
-  allExistingFiles.forEach(existingFile => {
+  allExistingFiles.forEach((existingFile) => {
     if (!map.has(path.normalize(existingFile))) {
       fs.unlinkSync(existingFile);
     }
@@ -207,18 +209,18 @@ function generate(
   );
 
   const generateNM2 = createNM2Generator({namespace: argv.namespace});
-  const generatorPropsH = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsH')
-    .generate;
-  const generatorPropsCPP = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsCPP')
-    .generate;
-  const generatorShadowNodeH = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeH')
-    .generate;
-  const generatorShadowNodeCPP = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeCPP')
-    .generate;
-  const generatorComponentDescriptorH = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateComponentDescriptorH')
-    .generate;
-  const generatorEventEmitterH = require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateEventEmitterH')
-    .generate;
+  const generatorPropsH =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsH').generate;
+  const generatorPropsCPP =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GeneratePropsCPP').generate;
+  const generatorShadowNodeH =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeH').generate;
+  const generatorShadowNodeCPP =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateShadowNodeCPP').generate;
+  const generatorComponentDescriptorH =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateComponentDescriptorH').generate;
+  const generatorEventEmitterH =
+    require('react-native-tscodegen/lib/rncodegen/src/generators/components/GenerateEventEmitterH').generate;
 
   normalizeFileMap(
     generateNM2(libraryName, schema, moduleSpecName),
@@ -228,7 +230,7 @@ function generate(
 
   if (
     Object.keys(schema.modules).some(
-      moduleName => schema.modules[moduleName].type === 'Component',
+      (moduleName) => schema.modules[moduleName].type === 'Component',
     )
   ) {
     const componentGenerators = [
@@ -240,7 +242,7 @@ function generate(
       generatorEventEmitterH,
     ];
 
-    componentGenerators.forEach(generator => {
+    componentGenerators.forEach((generator) => {
       const generated: Map<string, string> = generator(
         libraryName,
         schema,