+++ packages/@react-native-windows/telemetry/src/telemetry.ts
@@ -35,7 +35,22 @@ interface CommandInfo {
 }
 
 // This is our key with the AI backend
-const DefaultSetupString = '795006ca-cf54-40ee-8bc6-03deb91401c3';
+const RNWSetupString = '795006ca-cf54-40ee-8bc6-03deb91401c3';
+
+// Environment variable to override the default setup string
+const ENV_SETUP_OVERRIDE = 'RNW_TELEMETRY_SETUP';
+
+// Environment variable to override the http proxy (such as http://localhost:8888 for Fiddler debugging)
+const ENV_PROXY_OVERRIDE = 'RNW_TELEMETRY_PROXY';
+
+export const CommandEventName = 'RNWCLI.Command';
+export const CodedErrorEventName = 'RNWCLI.CodedError';
+
+// These are the event names we're tracking
+export const EventNamesWeTrack: string[] = [
+  CommandEventName,
+  CodedErrorEventName,
+];
 
 // These are NPM packages we care about, in terms of capturing versions used
 // and getting more details about when reporting errors
@@ -62,7 +77,7 @@ export const NuGetPackagesWeTrack: string[] = [
 export class Telemetry {
   protected static client?: appInsights.TelemetryClient = undefined;
   protected static options: TelemetryOptions = {
-    setupString: DefaultSetupString, // We default to our AI key, but callers can easily override it in setup
+    setupString: Telemetry.getDefaultSetupString(), // We default to our AI key, but callers can easily override it in setup
     preserveErrorMessages: false,
   };
 
@@ -73,6 +88,11 @@ export class Telemetry {
     | projectUtils.AppProjectInfo
     | projectUtils.DependencyProjectInfo = undefined;
 
+  protected static getDefaultSetupString(): string {
+    // Enable overriding the default setup string via an environment variable
+    return process.env[ENV_SETUP_OVERRIDE] ?? RNWSetupString;
+  }
+
   protected static reset(): void {
     // Reset client
     if (Telemetry.client) {
@@ -82,7 +102,7 @@ export class Telemetry {
 
     // Reset local members
     Telemetry.options = {
-      setupString: DefaultSetupString,
+      setupString: Telemetry.getDefaultSetupString(),
       preserveErrorMessages: false,
     };
     Telemetry.commandInfo = {};
@@ -131,9 +151,12 @@ export class Telemetry {
       Telemetry.options.setupString,
     );
 
-    // Uncomment for Fiddler testing
-    // Telemetry.client.config.proxyHttpUrl = 'http://localhost:8888';
-    // Telemetry.client.config.proxyHttpsUrl = 'http://localhost:8888';
+    // Allow overriding the proxy server via an environment variable
+    const proxyServer = process.env[ENV_PROXY_OVERRIDE];
+    if (proxyServer) {
+      Telemetry.client.config.proxyHttpUrl = proxyServer;
+      Telemetry.client.config.proxyHttpsUrl = proxyServer;
+    }
 
     Telemetry.client.config.disableAppInsights = Telemetry.isTest;
     Telemetry.client.channel.setUseDiskRetryCaching(!Telemetry.isTest);
@@ -141,8 +164,10 @@ export class Telemetry {
 
   /** Sets up any base properties that all telemetry events require. */
   private static async setupBaseProperties() {
-    Telemetry.client!.commonProperties.deviceId = await basePropUtils.deviceId();
-    Telemetry.client!.commonProperties.deviceLocale = await basePropUtils.deviceLocale();
+    Telemetry.client!.commonProperties.deviceId =
+      await basePropUtils.deviceId();
+    Telemetry.client!.commonProperties.deviceLocale =
+      await basePropUtils.deviceLocale();
     Telemetry.client!.commonProperties.deviceNumCPUs = basePropUtils
       .deviceNumCPUs()
       .toString();
@@ -191,7 +216,17 @@ export class Telemetry {
     },
   ): boolean {
     delete envelope.tags['ai.cloud.roleInstance'];
-    return true;
+
+    // Filter out "legacy" events from older stable branches
+    const properties = envelope.data.baseData?.properties;
+    if (
+      properties?.eventName &&
+      EventNamesWeTrack.includes(properties.eventName)
+    ) {
+      return true;
+    }
+
+    return false;
   }
 
   /**
@@ -347,7 +382,7 @@ export class Telemetry {
 
   private static trackCommandEvent(extraProps?: Record<string, any>) {
     const props: Record<string, any> = {
-      eventName: 'RNWCLI.Command',
+      eventName: CommandEventName,
     };
 
     // Set command props
@@ -377,7 +412,7 @@ export class Telemetry {
     }
 
     const props: Record<string, any> = {
-      eventName: 'RNWCLI.CodedError',
+      eventName: CodedErrorEventName,
     };
 
     // Save off CodedError info