+++ vnext/src/Libraries/Image/Image.windows.js
@@ -36,7 +36,7 @@ function getSize(
     .then(([width, height]) => success(width, height))
     .catch(
       failure ||
-        function() {
+        function () {
           console.warn('Failed to get size for image ' + uri);
         },
     );
@@ -49,12 +49,12 @@ function getSizeWithHeaders(
   failure?: (error: any) => void,
 ): any {
   return NativeImageLoaderIOS.getSizeWithHeaders(uri, headers)
-    .then(function(sizes) {
+    .then(function (sizes) {
       success(sizes.width, sizes.height);
     })
     .catch(
       failure ||
-        function() {
+        function () {
           console.warn('Failed to get size for image: ' + uri);
         },
     );
@@ -103,7 +103,7 @@ type ImageComponentStatics = $ReadOnly<{|
  * including network images, static resources, temporary local images, and
  * images from local disk, such as the camera roll.
  *
- * See https://reactnative.dev/docs/image.html
+ * See https://reactnative.dev/docs/image
  */
 let Image = (props: ImagePropsType, forwardedRef) => {
   const source = resolveAssetSource(props.source) || {
@@ -145,7 +145,7 @@ let Image = (props: ImagePropsType, forwardedRef) => {
   return (
     // [Windows
     <TextAncestor.Consumer>
-      {hasTextAncestor => {
+      {(hasTextAncestor) => {
         invariant(
           !hasTextAncestor,
           'Nesting of <Image> within <Text> is not currently supported.',
@@ -154,7 +154,7 @@ let Image = (props: ImagePropsType, forwardedRef) => {
 
         return (
           <ImageAnalyticsTagContext.Consumer>
-            {analyticTag => {
+            {(analyticTag) => {
               return (
                 <ImageViewNativeComponent
                   {...props}
@@ -188,7 +188,7 @@ Image.displayName = 'Image';
 /**
  * Retrieve the width and height (in pixels) of an image prior to displaying it.
  *
- * See https://reactnative.dev/docs/image.html#getsize
+ * See https://reactnative.dev/docs/image#getsize
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,
@@ -199,7 +199,7 @@ Image.getSize = getSize;
  * Retrieve the width and height (in pixels) of an image prior to displaying it
  * with the ability to provide the headers for the request.
  *
- * See https://reactnative.dev/docs/image.html#getsizewithheaders
+ * See https://reactnative.dev/docs/image#getsizewithheaders
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,
@@ -210,7 +210,7 @@ Image.getSizeWithHeaders = getSizeWithHeaders;
  * Prefetches a remote image for later use by downloading it to the disk
  * cache.
  *
- * See https://reactnative.dev/docs/image.html#prefetch
+ * See https://reactnative.dev/docs/image#prefetch
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,
@@ -221,7 +221,7 @@ Image.prefetch = prefetch;
  * Prefetches a remote image for later use by downloading it to the disk
  * cache, and adds metadata for queryRootName and rootTag.
  *
- * See https://reactnative.dev/docs/image.html#prefetch
+ * See https://reactnative.dev/docs/image#prefetch
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,
@@ -231,7 +231,7 @@ Image.prefetchWithMetadata = prefetchWithMetadata;
 /**
  * Performs cache interrogation.
  *
- *  See https://reactnative.dev/docs/image.html#querycache
+ *  See https://reactnative.dev/docs/image#querycache
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,
@@ -241,7 +241,7 @@ Image.queryCache = queryCache;
 /**
  * Resolves an asset reference into an object.
  *
- * See https://reactnative.dev/docs/image.html#resolveassetsource
+ * See https://reactnative.dev/docs/image#resolveassetsource
  */
 /* $FlowFixMe[prop-missing] (>=0.89.0 site=react_native_ios_fb) This comment
  * suppresses an error found when Flow v0.89 was deployed. To see the error,