+++ vnext/src/Libraries/Network/RCTNetworkingWinShared.js
@@ -9,8 +9,8 @@
 'use strict';
 
 import RCTDeviceEventEmitter from '../EventEmitter/RCTDeviceEventEmitter';
-const RCTNetworkingNative = require('../BatchedBridge/NativeModules')
-  .Networking; // [Windows]
+const RCTNetworkingNative =
+  require('../BatchedBridge/NativeModules').Networking; // [Windows]
 import {type NativeResponseType} from './XMLHttpRequest';
 import convertRequestBody, {type RequestBody} from './convertRequestBody';
 import {type EventSubscription} from '../vendor/emitter/EventEmitter';