+++ packages/@rnw-scripts/stamp-version/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @rnw-scripts/stamp-version
 
-This log was last generated on Thu, 18 Nov 2021 06:11:21 GMT and should not be manually modified.
+This log was last generated on Fri, 03 Dec 2021 06:12:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0
+
+Fri, 03 Dec 2021 06:12:11 GMT
+
+### Patches
+
+- Bump @react-native-windows/fs to v1.0.1
+- Bump @react-native-windows/package-utils to v0.0.0-canary.24
+- Bump @rnw-scripts/eslint-config to v1.1.10
+- Bump @rnw-scripts/just-task to v2.2.2
 
 ## 0.0.0