+++ packages/sample-apps/CHANGELOG.md
@@ -1,8 +1,118 @@
 # Change Log - sample-apps
 
-This log was last generated on Thu, 18 Nov 2021 06:11:21 GMT and should not be manually modified.
+This log was last generated on Tue, 28 Dec 2021 06:09:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0
+
+Tue, 28 Dec 2021 06:09:21 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.433
+
+## 0.0.0
+
+Sat, 18 Dec 2021 06:08:07 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.432
+
+## 0.0.0
+
+Thu, 16 Dec 2021 06:11:19 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.431
+- Bump @react-native-windows/virtualized-list to v0.0.0-canary.22
+
+## 0.0.0
+
+Tue, 14 Dec 2021 06:12:27 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.430
+
+## 0.0.0
+
+Fri, 10 Dec 2021 06:11:42 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.429
+- Bump @react-native-windows/virtualized-list to v0.0.0-canary.21
+
+## 0.0.0
+
+Tue, 07 Dec 2021 06:10:13 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.428
+
+## 0.0.0
+
+Sat, 04 Dec 2021 06:09:39 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.426
+
+## 0.0.0
+
+Fri, 03 Dec 2021 06:12:11 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.425
+- Bump @react-native-windows/codegen to v0.0.0-canary.22
+- Bump @react-native-windows/virtualized-list to v0.0.0-canary.20
+- Bump @rnw-scripts/eslint-config to v1.1.10
+- Bump @rnw-scripts/just-task to v2.2.2
+
+## 0.0.0
+
+Thu, 02 Dec 2021 06:09:50 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.424
+
+## 0.0.0
+
+Fri, 26 Nov 2021 06:09:54 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.423
+
+## 0.0.0
+
+Thu, 25 Nov 2021 06:09:25 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.422
+
+## 0.0.0
+
+Wed, 24 Nov 2021 06:08:49 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.421
+
+## 0.0.0
+
+Tue, 23 Nov 2021 06:09:26 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.420
 
 ## 0.0.0